big layout change. change app.py to be use new libs
