+++ ArduinoComms/ArduinoComm.py
@@ -5,6 +5,8 @@ import time
 class ArduinoComm:
 	def __init__(self):
 		self.arduinoSerial = serial.Serial('com6', 9600, timeout=1) #opens comport
+		time.sleep(5) # wait until device is setup 
+		print("arduino comms set up!")
 
 	def write_read (self, x : str): 
 		bytestr = bytes(x, 'utf-8')
@@ -13,7 +15,6 @@ class ArduinoComm:
 		return data.decode().strip()
 
 if __name__ == "__main__": # enables us to import the function but not run the main code, as well as running the main code when we want to test. see https://stackoverflow.com/questions/419163/what-does-if-name-main-do
-	time.sleep(5) # wait until device is setup 
 	arduino = ArduinoComm()
 	while True: 
 		num = input("Enter a number: ") # Taking input from user
